refactor signed_entity service
