+++ mithril-aggregator/src/dependency_injection/builder.rs
@@ -51,7 +51,7 @@ use crate::{
     },
     event_store::{EventMessage, EventStore, TransmitterService},
     http_server::routes::router,
-    signed_entity_service::{MithrilSignedEntityService, SignedEntityService},
+    services::{MithrilSignedEntityService, SignedEntityService},
     signer_registerer::SignerRecorder,
     stake_distribution_service::{MithrilStakeDistributionService, StakeDistributionService},
     ticker_service::{MithrilTickerService, TickerService},
