+++ mithril-aggregator/src/http_server/routes/middlewares.rs
@@ -1,6 +1,6 @@
 use crate::certifier_service::CertifierService;
 use crate::event_store::{EventMessage, TransmitterService};
-use crate::signed_entity_service::SignedEntityService;
+use crate::services::SignedEntityService;
 use crate::ticker_service::TickerService;
 use crate::{
     dependency_injection::MultiSignerWrapper, CertificatePendingStore, Configuration,
