+++ mithril-aggregator/src/services/signed_entity.rs
